Fixed some warnings in atom1.5, added canDaemon emulator
git-svn-id: http://svn.arune.se/svn/HomeAutomation/trunk@1914 e69ec2e4-db17-0410-bd4a-d6f28f89cd0d

+++ PcSoftware/Atom1.5/src/atomic.cpp
@@ -335,7 +335,7 @@ static char** AutoComplete(const char* text, int start, int end)
     unsigned int count = 0;
     boost::mutex guard_mutex;
 
-    for (unsigned int n = 0; n < start; n++)
+    for (unsigned int n = 0; n < (unsigned int)start; n++)
     {
         if (rl_line_buffer[n] == ' ')
         {
@@ -367,7 +367,7 @@ char* AutoCompleteGet(const char* text, int state)
         index = 0;
     }
     
-    while (autocomplete_list.size() > index)
+    while (autocomplete_list.size() > (unsigned int) index)
     {
         name = autocomplete_list[index];
